Move Featured RecyclerView to own file
