+++ app/src/main/java/com/lambda_labs/community_calendar/adapter/FeaturedRecyler.kt
@@ -1,0 +1,33 @@
+package com.lambda_labs.community_calendar.adapter
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.textview.MaterialTextView
+import com.lambda_labs.community_calendar.R
+import kotlinx.android.synthetic.main.featured_event_recycler_item.view.*
+
+//    Recycler for Featured Events
+class FeaturedEventRecycler(private val events: ArrayList<String>) :
+    RecyclerView.Adapter<FeaturedEventRecycler.ViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val view = LayoutInflater.from(parent.context)
+            .inflate(R.layout.featured_event_recycler_item, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun getItemCount(): Int = events.size
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        val event = events[position]
+
+        holder.title.text = event
+    }
+
+    inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+        val title: MaterialTextView = view.txt_featured_name
+    }
+}
+
