+++ app/src/main/java/com/lambda_labs/community_calendar/view/HomeFragment.kt
@@ -16,6 +16,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.textview.MaterialTextView
 import com.lambda_labs.community_calendar.R
 import com.lambda_labs.community_calendar.adapter.EventRecycler
+import com.lambda_labs.community_calendar.adapter.FeaturedEventRecycler
 import com.lambda_labs.community_calendar.util.*
 import com.lambda_labs.community_calendar.viewmodel.HomeViewModel
 import kotlinx.android.synthetic.main.featured_event_recycler_item.view.*
@@ -185,31 +186,4 @@ class HomeFragment : Fragment() {
         }
 
     }
-
-
-    //    Recycler for Featured Events
-    inner class FeaturedEventRecycler(private val events: ArrayList<String>) :
-        RecyclerView.Adapter<FeaturedEventRecycler.ViewHolder>() {
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            val view = LayoutInflater.from(parent.context)
-                .inflate(R.layout.featured_event_recycler_item, parent, false)
-            return ViewHolder(view)
-        }
-
-        override fun getItemCount(): Int = events.size
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val event = events[position]
-
-            holder.title.text = event
-        }
-
-        inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-            val title: MaterialTextView = view.txt_featured_name
-        }
-    }
-
-
-
 }
